+++ src/Exceptions/RuleException.php
@@ -18,11 +18,6 @@ class RuleException extends Exception
         return new self(sprintf('Validator "%s" is not registered', $rule));
     }
 
-    public static function cannotOverrideExistingRule(string $rule): self
-    {
-        return new self(sprintf('A rule named "%s" is already registered and cannot be overridden. Specify a different name for your rule', $rule));
-    }
-
     public static function invalidRuleType(string $rule): self
     {
         return new self(sprintf('Rule must be a string, Closure or "%s" instance; "%s" given', Rule::class, $rule));