+++ CHANGELOG.md
@@ -1,7 +1,18 @@
 Change Log
 ==========
 
+2021-11-09
+----------
+
+ * remove the humanize keys and override checks - not needed
+ * add support for aliases on rules and drop pre-process of input data
+ * add support for rules as array of rule -> value as well as [ rule:value ]
+ * correct usage of assertEquals in tests
+
 2021-11-04
 ----------
 
-Initial commit with ideas from https://github.com/rakit/validation
+ * initial commit porting code from https://github.com/rakit/validation
+ * clean up some internals
+ * make use of PHP8 features
+ * rename some classes